+++ messaging/src/main/resources/org/jboss/as/messaging/LocalDescriptions.properties
@@ -436,8 +436,19 @@ jms-queue.consumer-count=The number of consumers consuming messages from this qu
 jms-queue.dead-letter-address=The address to send dead messages to.
 jms-queue.delivering-count=The number of messages that this queue is currently delivering to its consumers.
 jms-queue.durable=Whether the queue is durable or not.
+jms-queue.elements.JMSPriority=The message's priority.
+jms-queue.elements.JMSTimestamp=The message's timestamp, in ms since the epoch.
+jms-queue.elements.JMSExpiration=The message's expiration time.
+jms-queue.elements.JMSDeliveryMode=The delivery mode for the message, PERSISTENT or NON_PERSISTENT.
+jms-queue.elements.JMSMessageID=The message's id.
 jms-queue.entries=The jndi names the queue will be bound to.
 jms-queue.expiry-address=The address to send expired messages to.
+jms-queue.list-delivering-messages.consumerName=The name of the consumer that the messages are delivered to.
+jms-queue.list-delivering-messages.elements=The delivering messages.
+jms-queue.list-delivering-messages.reply=A list where each element is a map of message properties to their value.
+jms-queue.list-delivering-messages=List the messages are currently being delivered.
+jms-queue.list-delivering-messages-as-json.reply=A string in JSON format representing a list where each element is a map of message properties to their value.
+jms-queue.list-delivering-messages-as-json=List the messages are currently being delivered as a string in JSON format
 jms-queue.list-messages.JMSDeliveryMode=The delivery mode for the message, PERSISTENT or NON_PERSISTENT.
 jms-queue.list-messages.JMSExpiration=The message's expiration time.
 jms-queue.list-messages.JMSMessageID=The message ID.
@@ -445,6 +456,15 @@ jms-queue.list-messages.JMSPriority=The message's priority.
 jms-queue.list-messages.JMSTimestamp=The message's timestamp, in ms since the epoch.
 jms-queue.list-messages.filter=A message filter. An undefined or empty filter will match all messages.
 jms-queue.list-messages=List all messages in the queue that match the filter
+jms-queue.list-scheduled-messages-as-json.reply=A string in JSON format representing a list where each element is a map of message properties to their value.
+jms-queue.list-scheduled-messages-as-json=List the messages scheduled for delivery as a string in JSON format.
+jms-queue.list-scheduled-messages.JMSDeliveryMode=The delivery mode for the message, PERSISTENT or NON_PERSISTENT.
+jms-queue.list-scheduled-messages.JMSExpiration=The message's expiration time.
+jms-queue.list-scheduled-messages.JMSMessageID=The message ID.
+jms-queue.list-scheduled-messages.JMSPriority=The message's priority.
+jms-queue.list-scheduled-messages.JMSTimestamp=The message's timestamp, in ms since the epoch.
+jms-queue.list-scheduled-messages.reply=A list where each element is a map of message properties to their value.
+jms-queue.list-scheduled-messages=List the messages scheduled for delivery.
 jms-queue.message-count=The number of messages currently in this queue.
 jms-queue.message.JMSDeliveryMode=The delivery mode for the message, PERSISTENT or NON_PERSISTENT.
 jms-queue.message.JMSExpiration=The message's expiration time.
